requirements.txt
transformers 
Pillow
ultralytics
